+++ src/main/java/de/alpharogroup/crypto/factories/KeyStoreFactory.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -24,12 +24,7 @@
  */
 package de.alpharogroup.crypto.factories;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
+import java.io.*;
 import java.security.KeyStore;
 import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
@@ -40,6 +35,34 @@ import java.security.cert.CertificateException;
  */
 public final class KeyStoreFactory
 {
+
+	/**
+	 * Factory method for create a new {@link KeyStore} object loaded from an existing
+	 * {@link KeyStore} object from the given file.
+	 *
+	 * @param type
+	 *            the type of the keystore
+	 * @param password
+	 *            the password of the keystore
+	 * @param keystoreFile
+	 *            the keystore file
+	 * @return the loaded {@link KeyStore} object
+	 * @throws KeyStoreException
+	 *             is thrown if there is an error accessing the key store
+	 * @throws NoSuchAlgorithmException
+	 *             is thrown if instantiation of the SecretKeyFactory object fails
+	 * @throws CertificateException
+	 *             is thrown if there is an error with an certificate
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static KeyStore loadKeyStore(final File keystoreFile, final String type, final String password)
+		throws KeyStoreException, NoSuchAlgorithmException, CertificateException, IOException
+	{
+		final KeyStore keyStore = KeyStore.getInstance(type);
+		keyStore.load(new FileInputStream(keystoreFile), password.toCharArray());
+		return keyStore;
+	}
 
 	/**
 	 * Factory method for load the {@link KeyStore} object from the given already existing file
@@ -61,10 +84,13 @@ public final class KeyStoreFactory
 	 *             Signals that an I/O exception has occurred.
 	 * @throws KeyStoreException
 	 *             is thrown if there is an error accessing the key store
+	 * @deprecated use instead the method <code>loadKeyStore</code>
+	 * Note: will be removed in the next minor version
 	 */
 	public static KeyStore newKeyStore(final String type, final String password,
-		final File keystoreFile) throws NoSuchAlgorithmException, CertificateException,
-		FileNotFoundException, IOException, KeyStoreException
+		final File keystoreFile)
+		throws NoSuchAlgorithmException, CertificateException, FileNotFoundException, IOException,
+		KeyStoreException
 	{
 		return newKeyStore(type, password, keystoreFile, false);
 	}
@@ -82,6 +108,8 @@ public final class KeyStoreFactory
 	 * @param createNewKeyStore
 	 *            if the {@linkplain KeyStore} should be new created.
 	 * @return the loaded {@link KeyStore} object
+	 * @throws KeyStoreException
+	 *             is thrown if there is an error accessing the key store
 	 * @throws NoSuchAlgorithmException
 	 *             is thrown if instantiation of the SecretKeyFactory object fails
 	 * @throws CertificateException
@@ -90,26 +118,50 @@ public final class KeyStoreFactory
 	 *             is thrown if the keystore file not found
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
-	 * @throws KeyStoreException
-	 *             is thrown if there is an error accessing the key store
 	 */
 	public static KeyStore newKeyStore(final String type, final String password,
-		final File keystoreFile, final boolean createNewKeyStore) throws NoSuchAlgorithmException,
-		CertificateException, FileNotFoundException, IOException, KeyStoreException
+		final File keystoreFile, final boolean createNewKeyStore)
+		throws KeyStoreException, NoSuchAlgorithmException, CertificateException,
+		FileNotFoundException, IOException
+	{
+		return createNewKeyStore ?
+			newKeyStore(keystoreFile, type, password) :
+			loadKeyStore(keystoreFile, type, password);
+	}
+
+	/**
+	 * Factory method for create a new empty {@link KeyStore} object and save it to the given file
+	 * with the given parameters.
+	 *
+	 * @param keystoreFile
+	 *            the keystore file
+	 * @param type
+	 *            the type of the keystore
+	 * @param password
+	 *            the password of the keystore
+	 * @return the loaded {@link KeyStore} object
+	 * @throws KeyStoreException
+	 *             is thrown if there is an error accessing the key store
+	 * @throws NoSuchAlgorithmException
+	 *             is thrown if instantiation of the SecretKeyFactory object fails
+	 * @throws CertificateException
+	 *             is thrown if there is an error with an certificate
+	 * @throws FileNotFoundException
+	 *             is thrown if the keystore file not found
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static KeyStore newKeyStore(final File keystoreFile, final String type,
+		final String password)
+		throws KeyStoreException, NoSuchAlgorithmException, CertificateException, IOException
 	{
 		final KeyStore keyStore = KeyStore.getInstance(type);
-		if (createNewKeyStore)
-		{
-			// initialize keystore
-			keyStore.load(null, password.toCharArray());
-			// get output stream
-			OutputStream out = new FileOutputStream(keystoreFile);
-			// store to new file
-			keyStore.store(out, password.toCharArray());
-			return keyStore;
-		}
-		// load from existing file
-		keyStore.load(new FileInputStream(keystoreFile), password.toCharArray());
+		// initialize keystore
+		keyStore.load(null, password.toCharArray());
+		// get output stream
+		OutputStream out = new FileOutputStream(keystoreFile);
+		// store to new file
+		keyStore.store(out, password.toCharArray());
 		return keyStore;
 	}